+++ packages/web3-eth/CHANGELOG.md
@@ -163,6 +163,11 @@ Documentation:
 
 ## [Unreleased]
 
+### Added
+
+-   A `rpc_method_wrapper` (`signTypedData`) for the rpc calls `eth_signTypedData` and `eth_signTypedData_v4` (#6286)
+-   A `signTypedData` method to the `Web3Eth` class (#6286)
+
 ### Fixed
 
 -   Missing `blockHeaderSchema` properties causing some properties to not appear in response of `newHeads` subscription (#6243)